Move send in desktop
