+++ apps/desktop/src/app/app.module.ts
@@ -47,12 +47,12 @@ import { AccountSwitcherComponent } from "./layout/account-switcher.component";
 import { HeaderComponent } from "./layout/header.component";
 import { NavComponent } from "./layout/nav.component";
 import { SearchComponent } from "./layout/search/search.component";
-import { AddEditComponent as SendAddEditComponent } from "./send/add-edit.component";
-import { EffluxDatesComponent as SendEffluxDatesComponent } from "./send/efflux-dates.component";
-import { SendComponent } from "./send/send.component";
 import { SharedModule } from "./shared/shared.module";
 import { GeneratorComponent } from "./tools/generator.component";
 import { PasswordGeneratorHistoryComponent } from "./tools/password-generator-history.component";
+import { AddEditComponent as SendAddEditComponent } from "./tools/send/add-edit.component";
+import { EffluxDatesComponent as SendEffluxDatesComponent } from "./tools/send/efflux-dates.component";
+import { SendComponent } from "./tools/send/send.component";
 import { ExportComponent } from "./vault/export.component";
 
 @NgModule({
